feat(app/): スクリーンセーバーをlayout.tsxに宣言して全ページで表示されるように
